In Kolab 3, the mail quota is displayed in KB. A quota of 2 GB becomes 2*1024*1024 = 2097152 This is quite difficult to read and to calculate as well.
The attached patch finds the closest unit that works without cutting something off (only use the next unit if modulo is 0).

The admin is presented a select box for KB, MB and GB.
The value is displayed either in KB, MB or GB.
When saving, the value is replaced again with the KB value, depending on the selected unit.

See: https://issues.kolab.org/show_bug.cgi?id=1966

Conflicts:
	doc/kolab_wap-3.0.sql
	doc/kolab_wap-3.1.sql